Move everything inside index.ios.js to components/App; then update index.ios.js and index.android.js to use App as the entry point (so now we can also use Android port of the app)
