+++ PuppyPlayDate/index.android.js
@@ -1,36 +1,4 @@
-/**
- * Sample React Native App
- * https://github.com/facebook/react-native
- */
+import { AppRegistry } from 'react-native';
+import App from './components/App';
 
-import React, { Component } from 'react';
-import {
-  AppRegistry,
-  StyleSheet,
-  Text,
-  View
-} from 'react-native';
-
-class PuppyPlayDate extends Component {
-  render() {
-    return (
-      <View style={styles.container}>
-        <Text style={styles.welcome}>
-          Welcome to React Native!
-        </Text>
-        <Text style={styles.instructions}>
-          To get started, edit index.android.js
-        </Text>
-        <Text style={styles.instructions}>
-          Shake or press menu button for dev menu
-        </Text>
-      </View>
-    );
-  }
-}
-
-const styles = require('./style.js')
-
-
-
-AppRegistry.registerComponent('PuppyPlayDate', () => PuppyPlayDate);
+AppRegistry.registerComponent('PuppyPlayDate', () => App);
